@@ -6,27 +6,27 @@
 
   <groupId>com.pj</groupId>
   <artifactId>browserutils</artifactId>
-  <version>1.0.0</version>
+  <version>2.0.0</version>
 
   <name>BrowserUtils</name>
   <url>https://github.com/pavankjadda/browserutils</url>
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <maven.compiler.source>11</maven.compiler.source>
-    <maven.compiler.target>11</maven.compiler.target>
-    <maven-compiler-plugin.version>3.8.1</maven-compiler-plugin.version>
-    <maven-deploy-plugin.version>2.8.2</maven-deploy-plugin.version>
+    <maven.compiler.source>17</maven.compiler.source>
+    <maven.compiler.target>17</maven.compiler.target>
+    <maven-compiler-plugin.version>3.12.1</maven-compiler-plugin.version>
+    <maven-deploy-plugin.version>3.1.1</maven-deploy-plugin.version>
     <javax.servlet-api.version>4.0.1</javax.servlet-api.version>
-    <junit.version>5.7.0</junit.version>
+    <jakarta.servlet-api.version>6.0.0</jakarta.servlet-api.version>
+    <junit.version>5.10.1</junit.version>
   </properties>
-
 
   <dependencies>
     <dependency>
-      <groupId>javax.servlet</groupId>
-      <artifactId>javax.servlet-api</artifactId>
-      <version>${javax.servlet-api.version}</version>
+      <groupId>jakarta.servlet</groupId>
+      <artifactId>jakarta.servlet-api</artifactId>
+      <version>${jakarta.servlet-api.version}</version>
     </dependency>
 
     <dependency>
@@ -36,7 +36,6 @@
       <scope>test</scope>
     </dependency>
   </dependencies>
-
 
   <build>
     <pluginManagement>
