@@ -7,37 +7,33 @@ package com.pj.constants;
  * @author Pavan Kumar Jadda
  * @since 2.0.0
  */
-public class Labels
-{
-	//Browser Labels
-	public static final String CHROME="Chrome";
-	public static final String CHROMIUM ="Chromium";
-	public static final String EDGE="Edge";
-	public static final String EDG="Edg";
-	public static final String FIREFOX="Firefox";
-	public static final String SAFARI="Safari";
-	public static final String OPERA="Opera";
-	public static final String TRIDENT="Trident";
-	public static final String MSIE="MSIE";
+public class Labels {
+    //Browser Labels
+    public static final String CHROME = "Chrome";
+    public static final String CHROMIUM = "Chromium";
+    public static final String EDGE = "Edge";
+    public static final String EDG = "Edg";
+    public static final String FIREFOX = "Firefox";
+    public static final String SAFARI = "Safari";
+    public static final String OPERA = "Opera";
+    public static final String TRIDENT = "Trident";
+    public static final String MSIE = "MSIE";
 
-	// OS Labels
-	public static final String UNKNOWN="Unknown";
-	public static final String WINDOWS="Windows";
-	public static final String WINDOWS_NT="Windows NT";
-	public static final String MAC_OS="Mac OS";
-	public static final String LINUX="Linux";
-	public static final String UNIX="Unix";
-	public static final String ANDROID="Android";
-	public static final String IOS="iOS";
+    // OS Labels
+    public static final String UNKNOWN = "Unknown";
+    public static final String WINDOWS = "Windows";
+    public static final String WINDOWS_NT = "Windows NT";
+    public static final String MAC_OS = "Mac OS";
+    public static final String LINUX = "Linux";
+    public static final String UNIX = "Unix";
+    public static final String ANDROID = "Android";
+    public static final String IOS = "iOS";
 
-	//Device labels
-	public static final String IPHONE="iPhone";
-	public static final String IPAD="iPad";
+    //Device labels
+    public static final String IPHONE = "iPhone";
+    public static final String IPAD = "iPad";
 
-
-
-	private Labels()
-	{
-		// Hides public constructor
-	}
+    private Labels() {
+        // Hides public constructor
+    }
 }
