@@ -9,13 +9,10 @@ import org.junit.jupiter.api.Test;
 /**
  * Unit test for simple App.
  */
-class BrowserNameDetectionTest
-{
-
+class BrowserNameDetectionTest {
     @Test
-    void testInternetExplorerDetection()
-    {
-        String userAgent="Mozilla/5.0 (Windows NT 10.0; WOW64; Trident/7.0; rv:11.0) like Gecko";
+    void testInternetExplorerDetection() {
+        String userAgent = "Mozilla/5.0 (Windows NT 10.0; WOW64; Trident/7.0; rv:11.0) like Gecko";
         System.out.println(BrowserUtil.getBrowserName(userAgent));
         System.out.println(BrowserUtil.getBrowserVersion(userAgent));
         System.out.println(OperatingSystemUtil.getOperatingSystemName(userAgent));
@@ -24,9 +21,8 @@ class BrowserNameDetectionTest
     }
 
     @Test
-    void testEdgeChromiumBrowserDetection()
-    {
-        String userAgent="Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/64.0.3282.140 Safari/537.36 Edge/18.17763";
+    void testEdgeChromiumBrowserDetection() {
+        String userAgent = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/64.0.3282.140 Safari/537.36 Edge/18.17763";
         System.out.println(BrowserUtil.getBrowserName(userAgent));
         System.out.println(BrowserUtil.getBrowserVersion(userAgent));
         System.out.println(OperatingSystemUtil.getOperatingSystemName(userAgent));
@@ -35,9 +31,8 @@ class BrowserNameDetectionTest
     }
 
     @Test
-    void testEdgeLegacyBrowserDetection()
-    {
-        String userAgent="Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/64.0.3282.140 Safari/537.36 Edg/18.17763";
+    void testEdgeLegacyBrowserDetection() {
+        String userAgent = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/64.0.3282.140 Safari/537.36 Edg/18.17763";
         System.out.println(BrowserUtil.getBrowserName(userAgent));
         System.out.println(BrowserUtil.getBrowserVersion(userAgent));
         System.out.println(OperatingSystemUtil.getOperatingSystemName(userAgent));
@@ -46,9 +41,8 @@ class BrowserNameDetectionTest
     }
 
     @Test
-    void testChromeBrowserDetection()
-    {
-        String userAgent="Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/89.0.4389.82 Safari/537.36";
+    void testChromeBrowserDetection() {
+        String userAgent = "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/89.0.4389.82 Safari/537.36";
         System.out.println(BrowserUtil.getBrowserName(userAgent));
         System.out.println(BrowserUtil.getBrowserVersion(userAgent));
         System.out.println(OperatingSystemUtil.getOperatingSystemName(userAgent));
@@ -57,8 +51,7 @@ class BrowserNameDetectionTest
     }
 
     @Test
-    void testOtherChromiumBrowserDetection()
-    {
+    void testOtherChromiumBrowserDetection() {
         String userAgent = "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/88.0.4324.152 Safari/537.36";
         System.out.println(BrowserUtil.getBrowserName(userAgent));
         System.out.println(BrowserUtil.getBrowserVersion(userAgent));
@@ -68,9 +61,8 @@ class BrowserNameDetectionTest
     }
 
     @Test
-    void testFirefoxBrowserDetection()
-    {
-        String userAgent="Mozilla/5.0 (Macintosh; Intel Mac OS X 10.15; rv:86.0) Gecko/20100101 Firefox/86.0";
+    void testFirefoxBrowserDetection() {
+        String userAgent = "Mozilla/5.0 (Macintosh; Intel Mac OS X 10.15; rv:86.0) Gecko/20100101 Firefox/86.0";
         System.out.println(BrowserUtil.getBrowserName(userAgent));
         System.out.println(BrowserUtil.getBrowserVersion(userAgent));
         System.out.println(OperatingSystemUtil.getOperatingSystemName(userAgent));
@@ -79,9 +71,8 @@ class BrowserNameDetectionTest
     }
 
     @Test
-    void testSafariBrowserDetection()
-    {
-        String userAgent="Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/14.0.3 Safari/605.1.15";
+    void testSafariBrowserDetection() {
+        String userAgent = "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/14.0.3 Safari/605.1.15";
         System.out.println(BrowserUtil.getBrowserName(userAgent));
         System.out.println(BrowserUtil.getBrowserVersion(userAgent));
         System.out.println(OperatingSystemUtil.getOperatingSystemName(userAgent));
@@ -90,9 +81,8 @@ class BrowserNameDetectionTest
     }
 
     @Test
-    void testOperaBrowserDetection()
-    {
-        String userAgent="Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/69.0.3497.100 Safari/537.36 OPR/56.0.3051.52";
+    void testOperaBrowserDetection() {
+        String userAgent = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/69.0.3497.100 Safari/537.36 OPR/56.0.3051.52";
         System.out.println(BrowserUtil.getBrowserName(userAgent));
         System.out.println(BrowserUtil.getBrowserVersion(userAgent));
         System.out.println(OperatingSystemUtil.getOperatingSystemName(userAgent));
